requirements.txt
flask
discord
pynacl
python-dotenv
waitress
DyslexicLogLog